[#31/mod]코드리뷰반영
